+++ app/src/main/java/woowacourse/shopping/data/source/OrderDataSource.kt
@@ -4,7 +4,7 @@ import retrofit2.Response
 import woowacourse.shopping.domain.model.Coupon
 
 interface OrderDataSource {
-    suspend fun orderItems(ids: List<Int>): Response<Unit>
+    suspend fun orderItems(ids: List<Long>): Response<Unit>
 
     suspend fun getCoupons(): Response<List<Coupon>>
 }
